Added schema inference for csv adaptor.
git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@2262 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 647448efd950afa83e8b4573264901808507fb9a

+++ lib/dbt_c++/streams.hpp
@@ -28,7 +28,7 @@ bool __verbose = false;
 std::ostream& operator<<(std::ostream &strm, const boost::any &a) {
 	try{
 		if( a.type() == typeid(int) )
-			return strm << any_cast<long>(a);
+			return strm << any_cast<int>(a);
 
 		else if( a.type() == typeid(long) )
             return strm << any_cast<long>(a);
